moving types to types pkg, getting snapshot
